Fix compile error. Looks like Java 1.5 auto converts between Integer and int. Unfortunately 1.4 doesn't.

+++ core/org.eclipse.cdt.core/src/org/eclipse/cdt/internal/errorparsers/ErrorPattern.java
@@ -84,7 +84,7 @@ public class ErrorPattern {
 	public int getLineNum(Matcher matcher) {
 		try {
 			return groupLineNum != 0
-				? Integer.valueOf(matcher.group(groupLineNum))
+				? Integer.valueOf(matcher.group(groupLineNum)).intValue()
 				: 0;
 		} catch (NumberFormatException e) {
 			return 0;
